Zmiany server 29.12

+++ routes/form.js
@@ -13,58 +13,48 @@ router.get("/", currentUser, async (req, res) => {
     }
 });
 
-/*
-//dodanie tego pozwalałoby userowi na edycje wizyty po jej zatwierdzeniu - user powinien jedynie anulować wizytę (delete) po jej zatwierdzeniu jeśli mu coś nie pasuje
-router.put("/:visitId", async (req, res) => {
+router.put("/status/:visitId", currentUser, async (req, res) => {
     try {
-        const visitId = req.params.visitId //pobiera visitId
-        const newData = req.body //pobiera dane z żądania do edycji
-        const { error } = validate(newData) //walidacja
-        if (error) {
-            return res.status(400).send({ message: error.details[0].message }) //jeśli błąd - zwraca jaki błąd
+        const visitId = req.params.visitId
+        const newStatus = req.body.status
+        // Pobranie wizyty z bazy danych na podstawie visitId
+        const visit = await Visit.findById(visitId)
+        if (!visit) {
+            return res.status(404).send({ message: "Visit not found" })
         }
-        const updatedVisit = await Visit.findByIdAndUpdate(visitId, req.body, { new: true }) //sprawdza czy istnieje wizyta z takim visitId i ją aktualizuje
-        if (!updatedVisit) {
-            return res.status(404).send({ message: "Visit not found" }) //jeśli nie - zwraca błąd
+        // Aktualizacja pola status, jeśli użytkownik ma odpowiednie uprawnienia
+        if (req.currentUser.role !== 'employee' && req.currentUser.role !== 'admin') {
+            return res.status(403).send({ message: "Access forbidden!" })
         }
-        res.status(200).send({ data: updatedVisit, message: "Visit updated successfully" }) //status ok - update danych
+        visit.status = newStatus
+        await visit.save()
+        return res.status(200).send({ data: visit, message: "Visit status updated successfully" })
     } catch (error) {
-        res.status(500).send({ message: "Internal Server Error" })
-        console.log(error)
+        res.status(500).send({ message: error.message })
     }
 });
-*/
-
-/*
-router.delete("/:visitId", async (req, res) => {  //!!!!!!!!!!!! do poprawy !!!!!!!!!!!!!
-    try {
-        const visitId = req.params.visitId; //pobiera visitId
-        const deletedVisit = await Visit.findByIdAndDelete(visitId) //sprawdza czy istnieje wizyta z visitId i usuwa ją
-        if (!deletedVisit) {
-            return res.status(404).send({ message: "Visit not found" }) //sprawdza czy wizyta została usunięta
-        }
-        res.status(200).send({ data: deletedVisit, message: "Visit deleted successfully" }) //jeśli usunięty zwraca sukces
-    } catch (error) {
-        res.status(500).send({ message: "Internal Server Error" })
-        console.log(error)
-    }
-});
-*/
 
 router.delete("/:visitId", currentUser, currentVisit, async (req, res) => {
     try {
         const visitId = req.params.visitId
         const currentDateTime = new Date() //aktualna data i czas
+        //console.log("req.currentUser._id:", req.currentUser._id);
+        //console.log("req.currentVisit.userId:", req.currentVisit.createdBy);
         //sprawdzamy, czy bieżąca wizyta należy do zalogowanego użytkownika
-        if (!req.currentVisit || !req.currentVisit.userId == req.currentUser._id) {
+        if (!req.currentVisit || req.currentVisit.createdBy.toString() !== req.currentUser._id.toString()) {
             return res.status(403).send({ message: "You are not authorized to delete this visit" })
         }
-        const visitDateTime = new Date(req.currentVisit.date + ' ' + req.currentVisit.time) //data i czas wizyty
-        const timeDiff = visitDateTime - currentDateTime //różnica czasu między obecnym a wizytą
+        const visitDate = new Date(req.currentVisit.date)
+        const visitTime = req.currentVisit.time
+        const visitDateTime = new Date(visitDate.getFullYear(), visitDate.getMonth(), visitDate.getDate(), parseInt(visitTime.split(":")[0]), parseInt(visitTime.split(":")[1]), 0)
+        const timeDiff = visitDateTime.getTime() - currentDateTime.getTime() // Różnica czasu między obecnym a wizytą w milisekundach
         const timeDiffInHours = timeDiff / (1000 * 60 * 60) //różnica czasu w godzinach
         const minimumHoursForCancellation = 24
-        if (timeDiffInHours <= minimumHoursForCancellation) {
-            return res.status(403).send({ message: "It's too late to cancel this visit" });
+        const isFutureVisit = visitDateTime > currentDateTime // Sprawdzamy, czy wizyta jest w przyszłości
+        console.log("currentDateTime", currentDateTime)
+        console.log("visitDateTime", visitDateTime)
+        if (!isFutureVisit || timeDiffInHours < minimumHoursForCancellation) {
+            return res.status(403).send({ message: "It's too late to cancel this visit" })
         }
         const deletedVisit = await Visit.findByIdAndDelete(visitId)
         if (!deletedVisit) {
@@ -77,5 +67,22 @@ router.delete("/:visitId", currentUser, currentVisit, async (req, res) => {
     }
 });
 
+router.delete("/forEmployeeOrAdmin/:visitId", async (req, res) => {
+    try {
+        const visitId = req.params.visitId
+        const visitToDelete = await Visit.findById(visitId)
+        if (!visitToDelete) {
+            return res.status(404).send({ message: "Visit not found" })
+        }
+        const deletedVisit = await Visit.findByIdAndDelete(visitId)
+        if (!deletedVisit) {
+            return res.status(404).send({ message: "Visit not found" })
+        }
+        res.status(200).send({ data: deletedVisit, message: "Visit deleted successfully" })
+    } catch (error) {
+        res.status(500).send({ message: "Internal Server Error" })
+        console.log(error)
+    }
+});
 
 module.exports = router;
